refactoring - Pull Up Method
