+++ base_pars_lib/async_camoufox_base_parser.py
@@ -1,6 +1,5 @@
 import asyncio
 import random
-from dataclasses import dataclass
 from typing import Any, Callable
 
 from camoufox.async_api import AsyncCamoufox
@@ -10,11 +9,6 @@ from base_pars_lib.config import logger
 from base_pars_lib.core.async_browsers_parser_base import AsyncBrowsersParserBase
 from base_pars_lib.exceptions.browser import BrowserIsNotInitError
 
-
-@dataclass
-class JsResponse:
-    text: str
-    url: str
 
 class AsyncCamoufoxBaseParser(AsyncBrowsersParserBase):
     def __init__(self) -> None:
@@ -28,13 +22,13 @@ class AsyncCamoufoxBaseParser(AsyncBrowsersParserBase):
         self.print_logs: bool = False
 
     async def _backoff_create_browser(
-        self,
-        proxy: list[dict] | dict[str, str] | None = None,
-        headless: bool | str = "virtual",
-        os: str = "linux",
-        geoip: bool = True,
-        increase_by_seconds: int = 10,
-        iter_count: int = 10,
+            self,
+            proxy: list[dict] | dict[str, str] | None = None,
+            headless: bool | str = "virtual",
+            os: str = "linux",
+            geoip: bool = True,
+            increase_by_seconds: int = 10,
+            iter_count: int = 10,
     ) -> Browser | None:
         """
         Создаёт браузер-менеджер и браузер
@@ -93,16 +87,16 @@ class AsyncCamoufoxBaseParser(AsyncBrowsersParserBase):
         return None
 
     async def _backoff_open_new_page(
-        self,
-        url: str,
-        is_page_loaded_check: Callable,
-        page: Page | None = None,
-        check_page: Callable = None,  # type: ignore[assignment]
-        check_page_args: dict | None = None,
-        load_timeout: int = 30,
-        increase_by_seconds: int = 10,
-        iter_count: int = 10,
-        **new_page_kwargs: Any,
+            self,
+            url: str,
+            is_page_loaded_check: Callable,
+            page: Page | None = None,
+            check_page: Callable = None,  # type: ignore[assignment]
+            check_page_args: dict | None = None,
+            load_timeout: int = 30,
+            increase_by_seconds: int = 10,
+            iter_count: int = 10,
+            **new_page_kwargs: Any,
     ) -> Page | None:
         """
         Открывает страницу по переданному url,
@@ -172,84 +166,3 @@ class AsyncCamoufoxBaseParser(AsyncBrowsersParserBase):
             await asyncio.sleep(i * increase_by_seconds)
 
         return None
-
-    async def _make_request_from_page(
-        self,
-        page: Page,
-        url: str | list[str],
-        method: str,
-        request_body: str | dict | list | None = None,
-        headers: str | dict | None = None,
-        log_request: bool = False,
-        return_response_object: bool = False,
-        iter_count: int = 3,
-    ) -> str | list[str] | JsResponse | list[JsResponse]:
-        """
-        Выполняет запрос через JS со страницы
-
-        :param page: Tab
-            Объект страницы
-        :param url: str | list[str]
-            Ссылка
-            Если передан список ссылок, запросы отправятся асинхронно
-        :param method: str
-            HTTP-метод
-        :param request_body: str | dict | None = None
-            Тело запроса
-        :param headers: str | dict | None = None
-            Хедеры запроса
-        :param log_request: bool = False
-            Вывод JS-кода запроса
-        :param return_response_object: bool = False
-            Если True — возвращает список объектов JsResponse(text, url)
-        :param iter_count: int = 3
-            Кол-во попыток
-        :return:
-            Текст с запрашиваемой страницы или объекты JsResponse
-        """
-
-        urls: list[str] = url if isinstance(url, list) else [url]
-
-        async def _safe_eval(one_url: str) -> str | None:
-            """
-            Выполняет JS-запрос через page.evaluate() с повторными попытками.
-
-            :param one_url: str
-                URL, на который выполняется запрос.
-            :return:
-                Текст ответа от JS-запроса или None, если после всех попыток не удалось
-                получить результат.
-            """
-            for i in range(1, iter_count + 1):
-                try:
-                    script = await self._make_js_script(
-                        url=one_url,
-                        method=method,
-                        request_body=request_body,
-                        headers=headers,
-                        log_request=log_request,
-                    )
-                    return await page.evaluate(script)
-                except Exception as Ex:
-                    if self.debug:
-                        logger.backoff_exception(Ex, url=one_url, iteration=i,
-                                                 print_logs=self.print_logs)
-            return None
-
-        responses: list[str | BaseException | None] = await asyncio.gather(
-            *[_safe_eval(one_url) for one_url in urls],
-            return_exceptions=True)  # type: ignore[return-value]
-
-        results = []
-        for resp, u in zip(responses, urls):
-            if isinstance(resp, str):
-                results.append(JsResponse(text=resp, url=u))
-            else:
-                results.append(JsResponse(text="", url=u))
-
-        if not return_response_object:
-            results = [r.text for r in results]
-
-        if isinstance(url, str):
-            return results[0]
-        return results
