+++ base_pars_lib/async_nodriver_base_parser.py
@@ -109,61 +109,6 @@ class AsyncNodriverBaseParser(AsyncBrowsersParserBase):
 
         return None
 
-    async def _make_request_from_page(
-            self,
-            page: Tab,
-            url: str | list[str],
-            method: str,
-            request_body: str | dict | None = None,
-            headers: str | dict | None = None,
-            log_request: bool = False,
-    ) -> str | list[str]:
-        """
-        Выполняет запрос через JS со страницы
-
-        :param page: Tab
-            Объект страницы
-        :param url: str | list[str]
-            Ссылка
-            Если передан список ссылок, запросы отправятся асинхронно
-        :param method: str
-            HTTP-метод
-        :param request_body: str | dict | None = None
-            Тело запроса
-        :param headers: str | dict | None = None
-            Хедеры запроса
-        :param log_request: bool = False
-            Вывод JS-кода запроса
-        :return:
-            Текст с запрашиваемой страницы
-        """
-
-        tasks: list = []
-        if isinstance(url, list):
-            for one_url in url:
-                script = await self._make_js_script(
-                    url=one_url,
-                    method=method,
-                    request_body=request_body,
-                    headers=headers,
-                    log_request=log_request,
-                )
-                tasks.append(page.evaluate(script, await_promise=True))
-        else:
-            script = await self._make_js_script(
-                url=url,
-                method=method,
-                request_body=request_body,
-                headers=headers,
-                log_request=log_request,
-            )
-            tasks.append(page.evaluate(script, await_promise=True))
-
-        responses = await asyncio.gather(*tasks)  # type: ignore[return-value]
-        if responses and len(responses) == 1 and not isinstance(url, list):
-            return responses[0]
-        return responses
-
     @staticmethod
     async def _make_chrome_proxy_extension(
             host: str,
